### What changes were proposed in this pull request?
introduce a new node `WindowGroupLimit` to filter out unnecessary rows based on rank computed on partial dataset.

it supports following pattern:
```
SELECT (... (row_number|rank|dense_rank)()
    OVER (
PARTITION BY ...
ORDER BY  ... ) AS rn)
WHERE rn (==|<|<=) k
        AND other conditions
```

For these three rank-like functions (row_number|rank|dense_rank), the rank of a key computed on partial dataset always <= its final rank computed on the whole dataset，so we can safely discard rows with partial rank > k, anywhere.

This PR also take over some functions from https://github.com/apache/spark/pull/34367.


### Why are the changes needed?

1. reduce the shuffle write
2. solve skewed-window problem.
3. improve the performance and TPC-DS.

### Does this PR introduce _any_ user-facing change?
'No'.
Just update the inner implementation and add a new config.


### How was this patch tested?

1. new test suites
2. new micro benchmark
```
[info] Benchmark Top-K:                                                 Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
[info] -----------------------------------------------------------------------------------------------------------------------------------------------
[info] ROW_NUMBER (PARTITION: , WindowGroupLimit: false)                        10869          11498         928          1.9         518.3       1.0X
[info] ROW_NUMBER (PARTITION: , WindowGroupLimit: true)                          3440           3700         201          6.1         164.0       3.2X
[info] ROW_NUMBER (PARTITION: PARTITION BY b, WindowGroupLimit: false)          23866          24130         230          0.9        1138.0       0.5X
[info] ROW_NUMBER (PARTITION: PARTITION BY b, WindowGroupLimit: true)            6436           6691         342          3.3         306.9       1.7X
[info] RANK (PARTITION: , WindowGroupLimit: false)                              11502          11756         250          1.8         548.4       0.9X
[info] RANK (PARTITION: , WindowGroupLimit: true)                                3424           3738         220          6.1         163.3       3.2X
[info] RANK (PARTITION: PARTITION BY b, WindowGroupLimit: false)                23780          25101         543          0.9        1133.9       0.5X
[info] RANK (PARTITION: PARTITION BY b, WindowGroupLimit: true)                  6508           6689         109          3.2         310.3       1.7X
[info] DENSE_RANK (PARTITION: , WindowGroupLimit: false)                        11711          11846         118          1.8         558.4       0.9X
[info] DENSE_RANK (PARTITION: , WindowGroupLimit: true)                          3359           3811         228          6.2         160.2       3.2X
[info] DENSE_RANK (PARTITION: PARTITION BY b, WindowGroupLimit: false)          23601          24712         484          0.9        1125.4       0.5X
[info] DENSE_RANK (PARTITION: PARTITION BY b, WindowGroupLimit: true)            6507           6627         108          3.2         310.3       1.7X
```

3. manual test on TPC-DS
TPC-DS data size: 2TB.
This improvement is valid for TPC-DS q67 and no regression for other test cases.

| TPC-DS Query   | Before(Seconds)  | After(Seconds)  | Speedup(Percent)  |
|  ----  | ----  | ----  | ----  |
| q67 | 997.7585 | 412.046 | 142.15% |
| All TPC-DS | 7076.6715 | 6472.0535 | 9.34% |